xiaohongshu-meme-analysis/requirements.txt
# Python依赖
# python>=3.10

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 文本处理和NLP
jieba>=0.42.1
snownlp>=0.12.3

# 爬虫相关
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
tavily-python>=0.3.0

# AI和LLM
openai>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
wordcloud>=1.9.2

# 机器学习
scikit-learn>=1.3.0

# 工具
python-dotenv>=1.0.0
tqdm>=4.66.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0
